added server health checkup and removed cyclic dependencies
