--- server/handlers/tests/admin_settings_test.go
@@ -1,1 +1,0 @@
-package tests
