--- server/handlers/tests/login_test.go
@@ -1,1 +1,0 @@
-package tests
