--- server/handlers/tests/mfa_test.go
@@ -1,1 +1,0 @@
-package tests
