--- server/handlers/tests/setup_test.go
@@ -1,1 +1,0 @@
-package tests
