--- server/handlers/tests/user_test.go
@@ -1,1 +1,0 @@
-package tests
