+++ server/models/device.go
@@ -2,8 +2,10 @@ package models
 
 import (
 	"context"
+	"github.com/google/uuid"
 	"github.com/mundanelizard/koyi/server/config"
 	"github.com/mundanelizard/koyi/server/helpers"
+	"net/http"
 	"time"
 )
 
@@ -46,4 +48,26 @@ func CountDevice(ctx context.Context, filter interface{}) (int64, error) {
 	collection := helpers.GetCollection(config.UserDatabaseName, deviceCollectionName)
 	count, err := collection.CountDocuments(ctx, filter)
 	return count, err
+}
+
+/**
+todo => extract and store
+tenant_id: the course creator school the visitor is checking
+raw: the raw ua
+type: desktop / mobile / tablet / bot / other
+browser_name
+browser_version
+os_name: Android / IOS / Windows / Mac
+os_version: OS Version
+hardware_details: hstore containing memory, processor, device_model, device_name
+connection_speed: hstore containing downlink_max, connection_type
+*/
+
+func ExtractDevice(r *http.Request, userId *string) *Device {
+	// todo => implement this
+	id := uuid.New().String()
+	return &Device{
+		ID:     &id,
+		UserId: userId,
+	}
 }
