+++ server/models/intent.go
@@ -67,3 +67,27 @@ func getVerificationIntentType(user *User) string {
 	}
 	return VerifyEmailIntent
 }
+
+var actionSubject = map[string]string{
+	VerifyEmailIntent:       "https://gmail.com/google.com",
+	VerifyPhoneNumberIntent: "https://email/intent/intent",
+}
+
+var actionHTML = map[string]string{
+	VerifyEmailIntent:       "https://gmail.com/google.com",
+	VerifyPhoneNumberIntent: "https://email/intent/intent",
+}
+
+var actionText = map[string]string{
+	VerifyEmailIntent:       "https://gmail.com/google.com",
+	VerifyPhoneNumberIntent: "https://email/intent/intent",
+}
+
+func GetEmailDetails(action string) (string, *string, *string) {
+	// todo => use a template file
+	text := actionText[action]
+	html := actionHTML[action]
+	subject := actionSubject[action]
+
+	return subject, &text, &html
+}
